refactor: convert homepage to server component with client extraction
